refactor(app): type graph-series FillAlpha as byte
[en] GraphSeriesDefinition.FillAlpha was an int cast to byte at the two
RateScopeRenderer use sites, so an alpha above 255 would silently wrap
instead of being caught. Type it as byte at the source: the 0-255 range is
now a compile-time guarantee and the redundant casts are dropped. No
behavior change for the existing values (0, 20); no runtime validation added.

[ko] GraphSeriesDefinition.FillAlpha가 int였고 RateScopeRenderer 두 사용처에서
byte로 캐스팅되어, 255를 넘는 알파 값이 잡히지 않고 조용히 wrap될 수 있었습니다.
소스에서 byte로 선언해 0-255 범위를 컴파일 타임에 보장하고 중복 캐스트를
제거합니다. 기존 값(0, 20)에 대한 동작 변화는 없으며, 런타임 검증은 추가하지
않았습니다.

+++ src/TimeGrapher.App/Rendering/RateScopeRenderer.cs
@@ -252,7 +252,7 @@ internal sealed class RateScopeRenderer
             if (spec.FillAlpha > 0)
             {
                 sc.FillY = true;
-                sc.FillYColor = Color.FromARGB(ThemeColor(spec)).WithAlpha((byte)spec.FillAlpha);
+                sc.FillYColor = Color.FromARGB(ThemeColor(spec)).WithAlpha(spec.FillAlpha);
             }
             sc.LegendText = spec.Name;
             _scopePlots.Add(sc);
@@ -284,7 +284,7 @@ internal sealed class RateScopeRenderer
             _scopePlots[i].LineColor = Color.FromARGB(color);
             if (_scopeSeries[i].FillAlpha > 0)
             {
-                _scopePlots[i].FillYColor = Color.FromARGB(color).WithAlpha((byte)_scopeSeries[i].FillAlpha);
+                _scopePlots[i].FillYColor = Color.FromARGB(color).WithAlpha(_scopeSeries[i].FillAlpha);
             }
         }
